requirements.txt
pydantic>=2.8,<3
email-validator>=2.2,<3
python-dateutil>=2.9,<3
redis>=5.0,<6
pymongo>=4.8,<5
structlog>=24.1,<25
pytest>=8.2,<9
pytest-cov>=5.0,<6
